requirements.txt
# Core Dependencies
streamlit>=1.31.0
google-generativeai>=0.8.0
pillow>=10.0.0

# Security
cryptography>=42.0.0

# Utilities
python-dateutil>=2.8.2
